requirements.txt
pandas==2.0.3
numpy==1.24.4
matplotlib==3.7.2
seaborn==0.12.2
prophet==1.1.4
streamlit==1.28.0
python-dotenv==1.0.0
openai==1.3.0
langchain-openai==0.0.5
langchain==0.0.335
